Refactors PWS to be able to fetch different types of HREFs
- Moves the ListPersonsOutputTranslator into its own service module for maintainability
- Updates Search service to call new PWS function
